Revert "add reviewers changes"
This reverts commit 22c82bea0caf3e0f29399100c1bb67b8003fc042.
