+++ test/pool/test_max_pool3d_bwd.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2024, Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2018-2023, Advanced Micro Devices, Inc. All rights reserved.
 
 #include "gtest/gtest.h"
-#include "profiler/profile_pool3d_max_bwd_impl.hpp"
+#include "profiler/profile_max_pool3d_bwd_impl.hpp"
 #include "test_pool_fwd_common.hpp"
 
 template <typename Tuple>
@@ -23,7 +23,7 @@ class TestMaxPool3dBwd : public ::testing::Test
         for(auto param : params)
         {
             bool success =
-                ck::profiler::profile_pool3d_max_bwd_impl<InDataType,
+                ck::profiler::profile_max_pool3d_bwd_impl<InDataType,
                                                           OutDataType,
                                                           IndexDataType,
                                                           DOutDataType,
@@ -43,9 +43,26 @@ class TestMaxPool3dBwd : public ::testing::Test
     }
 };
 
+#if defined(CK_ENABLE_FP16) && defined(CK_ENABLE_BF16) && defined(CK_ENABLE_FP32)
 using KernelTypes = ::testing::Types<std::tuple<F16, F16, I32, NDHWC, NDHWC>,
                                      std::tuple<BF16, BF16, I32, NDHWC, NDHWC>,
                                      std::tuple<F32, F32, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_FP16) && defined(CK_ENABLE_FP32)
+using KernelTypes = ::testing::Types<std::tuple<F16, F16, I32, NDHWC, NDHWC>,
+                                     std::tuple<F32, F32, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_BF16) && defined(CK_ENABLE_FP32)
+using KernelTypes = ::testing::Types<std::tuple<BF16, BF16, I32, NDHWC, NDHWC>,
+                                     std::tuple<F32, F32, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_FP16) && defined(CK_ENABLE_BF16)
+using KernelTypes = ::testing::Types<std::tuple<F16, F16, I32, NDHWC, NDHWC>,
+                                     std::tuple<BF16, BF16, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_FP16)
+using KernelTypes = ::testing::Types<std::tuple<F16, F16, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_BF16)
+using KernelTypes = ::testing::Types<std::tuple<BF16, BF16, I32, NDHWC, NDHWC>>;
+#elif defined(CK_ENABLE_FP32)
+using KernelTypes = ::testing::Types<std::tuple<F32, F32, I32, NDHWC, NDHWC>>;
+#endif
 
 TYPED_TEST_SUITE(TestMaxPool3dBwd, KernelTypes);
 TYPED_TEST(TestMaxPool3dBwd, Test_Pool)
