+++ test/pool/test_max_pool3d_fwd.cpp
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2024, Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2018-2023, Advanced Micro Devices, Inc. All rights reserved.
 
 #include "gtest/gtest.h"
 #include "profiler/profile_pool3d_fwd_impl.hpp"
 #include "test_pool_fwd_common.hpp"
 
 template <typename Tuple>
-class TestAvgPool3dFwd : public ::testing::Test
+class TestMaxPool3dFwd : public ::testing::Test
 {
     protected:
     using InDataType      = std::tuple_element_t<0, Tuple>;
@@ -20,6 +20,7 @@ class TestAvgPool3dFwd : public ::testing::Test
     {
         for(auto param : params)
         {
+            // max pool
             bool success =
                 ck::profiler::profile_pool3d_fwd_impl<InDataType,
                                                       OutDataType,
@@ -27,7 +28,7 @@ class TestAvgPool3dFwd : public ::testing::Test
                                                       IndexDataType,
                                                       ck::tensor_layout::convolution::NDHWC,
                                                       ck::tensor_layout::convolution::NDHWC,
-                                                      ck::ReduceTensorOp::AVG,
+                                                      ck::ReduceTensorOp::MAX,
                                                       false,
                                                       false>(true,
                                                              2,
@@ -40,15 +41,40 @@ class TestAvgPool3dFwd : public ::testing::Test
                                                              param.input_left_pads_,
                                                              param.input_right_pads_);
             EXPECT_TRUE(success);
+
+            // max pool + index
+            success = ck::profiler::profile_pool3d_fwd_impl<InDataType,
+                                                            OutDataType,
+                                                            ComputeDataType,
+                                                            IndexDataType,
+                                                            ck::tensor_layout::convolution::NDHWC,
+                                                            ck::tensor_layout::convolution::NDHWC,
+                                                            ck::ReduceTensorOp::MAX,
+                                                            false,
+                                                            true>(true,
+                                                                  2,
+                                                                  false,
+                                                                  false,
+                                                                  param.length_,
+                                                                  param.window_spatial_lengths_,
+                                                                  param.window_strides_,
+                                                                  param.window_dilations_,
+                                                                  param.input_left_pads_,
+                                                                  param.input_right_pads_);
+            EXPECT_TRUE(success);
         }
     }
 };
 
+#ifdef CK_ENABLE_FP16
 using KernelTypes =
     ::testing::Types<std::tuple<F16, F16, F32, I32>, std::tuple<F32, F32, F32, I32>>;
+#else
+using KernelTypes = ::testing::Types<std::tuple<F32, F32, F32, I32>>;
+#endif
 
-TYPED_TEST_SUITE(TestAvgPool3dFwd, KernelTypes);
-TYPED_TEST(TestAvgPool3dFwd, Test_Pool)
+TYPED_TEST_SUITE(TestMaxPool3dFwd, KernelTypes);
+TYPED_TEST(TestMaxPool3dFwd, Test_Pool)
 {
     // length, window_length, window_stride, window_dilation, left_pad, right_pad
     this->params = {{{1, 1, 1, 1, 1}, {1, 1, 1}, {1, 1, 1}, {1, 1, 1}, {0, 0, 0}, {0, 0, 0}},
